rename errors module to result
