+++ src/storage/write_ahead_log.rs
@@ -1,4 +1,4 @@
-use crate::errors::{Errors, Result};
+use crate::result::{Errors, Result};
 use crate::options::DharmaOpts;
 use crate::storage::block::{create_blocks, write_block_to_disk, Block, Value};
 use crate::storage::sorted_string_table_reader::SSTableReader;
