+++ tests/persistence_test.rs
@@ -1,7 +1,7 @@
 use crate::common::test_key::TestKey;
 use crate::common::test_value::TestValue;
 use crate::common::{cleanup_paths, get_test_data, get_test_data_in_range};
-use dharmadb::errors::{Errors, Result};
+use dharmadb::result::{Errors, Result};
 use dharmadb::options::DharmaOpts;
 use dharmadb::persistence::Persistence;
 use dharmadb::storage::sorted_string_table_reader::SSTableReader;
